src/az-content-understanding/pyproject.toml
[project]
name = "az-content-understanding"
version = "0.1.0"
description = "Azure Content Understanding sample client"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "python-dotenv",
    "azure-identity>=1.23.0",
]

[tool.uv]
# uv-specific settings can be added here if needed
